Cleanup & move meteor method for parsing to helper function
