+++ dashboard/client/index.js
@@ -32,17 +32,7 @@ Template.dashboard.onCreated(function () {
   }
 
   instance.getChartData = opts => new Promise((resolve, reject) => {
-    Meteor.call('getAggregatedData', opts, (err, res) => {
-      if (err) reject(err)
-      Meteor.call('parseDataForNvd', res, (err1, res1) => {
-        if (err1) reject(err1)
-        resolve(res1)
-      })
-    })
-  })
-
-  instance.parseDataForNvd = data => new Promise((resolve, reject) => {
-    Meteor.call('parseDataForNvd', data, (err, res) => {
+    Meteor.call('getChartData', opts, (err, res) => {
       if (err) reject(err)
       resolve(res)
     })
